All: clean up

+++ Plugins/org.opendarts.prototype/src/main/java/org/opendarts/prototype/internal/service/stats/x01/entry/AvgEntry.java
@@ -84,7 +84,7 @@ public class AvgEntry implements Serializable {
 		if (this.count < 1) {
 			result = "-";
 		} else {
-			double d = this.sum / ((double) this.count);
+			double d = this.sum / this.count;
 			result = StatsValue.DOUBLE_FORMATTER.format(d);
 		}
 		return result;
